Move this to move data utils
